+++ main.go
@@ -3,7 +3,7 @@ Copyright © 2025 NAME HERE <EMAIL ADDRESS>
 */
 package main
 
-import "dotm/cmd"
+import "github.com/w31r4/dotm/cmd"
 
 func main() {
 	cmd.Execute()